+++ src/app/albaformList/types.ts
@@ -1,8 +1,5 @@
 import { SetStateAction } from 'react';
 
-
-
-
 export interface FilterContainerProps {
   isSort: 'mostRecent' | 'mostCommented' | 'mostLiked';
   setIsSort: React.Dispatch<
