+++ src/app/albatalk/types.ts
@@ -1,0 +1,8 @@
+import { SetStateAction } from "react";
+
+export interface FilterContainerProps {
+  isSort: 'mostRecent' | 'mostCommented' | 'mostLiked';
+  setIsSort: React.Dispatch<
+    SetStateAction<'mostRecent' | 'mostCommented' | 'mostLiked'>
+  >;
+}
