+++ src/app/mypage/owner/components/FilterContainer.tsx
@@ -1,33 +1,21 @@
-'use client';
-
-import {
-  SlideBg,
-  DropdownBox,
-  TabButton,
-  TabWrapper,
-  DropdownContainer,
-} from '../../styles';
-import Image from 'next/image';
-import SortDropdown from './SortDropdown';
-import { useClickOutside } from '@/hooks/common/useClickOutside';
+import { SlideBg, TabButton, TabWrapper } from '../../styles';
 import { FilterContainerProps } from '../../types';
+import PostSortButton from '@/components/postSort/PostSortButton';
 
 export default function FilterContainer({
   selectedTab,
   setSelectedTab,
-  isPostSort,
-  setIsPostSort,
+  isSort,
+  setIsSort,
 }: FilterContainerProps) {
-  const { outRef, dropdown, setDropdown } = useClickOutside();
-
   const handleTabClickPost = () => {
     setSelectedTab('post');
-    setIsPostSort('mostRecent');
+    setIsSort('mostRecent');
   };
 
   const handleTabClickComment = () => {
     setSelectedTab('comment');
-    setIsPostSort('mostRecent');
+    setIsSort('mostRecent');
   };
 
   return (
@@ -53,38 +41,7 @@ export default function FilterContainer({
           </TabWrapper>
         </div>
         {selectedTab === 'post' && (
-          <div className='relative' ref={outRef}>
-            <button
-              className='flex items-center'
-              onClick={() => setDropdown((prev) => !prev)}
-            >
-              <p className='pr-1 max-[768px]:text-[12px]'>
-                {isPostSort === 'mostRecent'
-                  ? '최신순'
-                  : isPostSort === 'mostCommented'
-                  ? '댓글순'
-                  : '좋아요순'}
-              </p>
-              <Image
-                src='/images/arrowDown.svg'
-                alt='arrow_down'
-                width={24}
-                height={24}
-                className={
-                  dropdown
-                    ? 'rotate-180 transition-transform duration-200'
-                    : 'transition-transform duration-200'
-                }
-              />
-            </button>
-            <DropdownContainer $active={dropdown}>
-              {dropdown && (
-                <DropdownBox>
-                  <SortDropdown isSort={isPostSort} setIsSort={setIsPostSort} />
-                </DropdownBox>
-              )}
-            </DropdownContainer>
-          </div>
+          <PostSortButton isSort={isSort} setIsSort={setIsSort} />
         )}
       </div>
     </>
