+++ src/components/postSort/PostSort.types.ts
@@ -1,0 +1,8 @@
+import { SetStateAction } from 'react';
+
+export interface SortDropdownProps {
+  isSort: 'mostRecent' | 'mostCommented' | 'mostLiked';
+  setIsSort: React.Dispatch<
+    SetStateAction<'mostRecent' | 'mostCommented' | 'mostLiked'>
+  >;
+}
